+++ mycodo/functions/display_ssd1306_oled_128x32_i2c_2lines.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 #
-#  display_ssd1306_oled_128x64_i2c.py - Function to output to LCD
+#  display_ssd1306_oled_128x32_i2c_2lines.py - Function to output to display
 #
 #  Copyright (C) 2015-2020 Kyle T. Gabriel <mycodo@kylegabriel.com>
 #
@@ -38,15 +38,15 @@ from mycodo.databases.models import FunctionChannel
 from mycodo.functions.base_function import AbstractFunction
 from mycodo.mycodo_flask.utils.utils_general import custom_channel_options_return_json
 from mycodo.mycodo_flask.utils.utils_general import delete_entry_with_id
-from mycodo.utils.constraints_pass import constraints_pass_positive_or_zero_value
-from mycodo.utils.constraints_pass import constraints_pass_positive_value
 from mycodo.utils.database import db_retrieve_table_daemon
 from mycodo.utils.functions import parse_function_information
 from mycodo.utils.lcd import format_measurement_line
 from mycodo.utils.system_pi import cmd_output
+from mycodo.utils.constraints_pass import constraints_pass_positive_or_zero_value
+from mycodo.utils.constraints_pass import constraints_pass_positive_value
 
 # Set to how many lines the LCD has
-lcd_lines = 8
+lcd_lines = 2
 
 
 def execute_at_creation(error, new_func, dict_functions=None):
@@ -164,13 +164,13 @@ def execute_at_modification(
 
 
 FUNCTION_INFORMATION = {
-    'function_name_unique': 'display_ssd1306_oled_128x64_i2c',
-    'function_name': 'Display: SSD1306 OLED 128x64 [8 Lines] (I2C)',
+    'function_name_unique': 'display_ssd1306_oled_128x32_i2c_2lines',
+    'function_name': 'Display: SSD1306 OLED 128x32 [2 Lines] (I2C)',
     'function_library': 'Adafruit-Circuitpython-SSD1306',
     'execute_at_creation': execute_at_creation,
     'execute_at_modification': execute_at_modification,
 
-    'message': 'This Function outputs to a 128x64 SSD1306 OLED display via I2C. This display Function will show 8 lines at a time, so channels are added in sets of 8 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next set of lines. Therefore, the first set of lines that are displayed are channels 0 - 7, then 8 - 15, and so on. After all channels have been displayed, it will cycle back to the beginning.',
+    'message': 'This Function outputs to a 128x32 SSD1306 OLED display via I2C. This display Function will show 2 lines at a time, so channels are added in sets of 2 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next set of lines. Therefore, the first set of lines that are displayed are channels 0 - 1, then 2 - 3, and so on. After all channels have been displayed, it will cycle back to the beginning.',
 
     'options_disabled': [
         'measurements_select',
@@ -234,7 +234,7 @@ FUNCTION_INFORMATION = {
         {
             'id': 'characters_x',
             'type': 'integer',
-            'default_value': 21,
+            'default_value': 17,
             'required': True,
             'name': 'Characters Per Line',
             'phrase': 'The maximum number of characters to display per line'
@@ -257,7 +257,7 @@ FUNCTION_INFORMATION = {
         {
             'id': 'font_size',
             'type': 'integer',
-            'default_value': 10,
+            'default_value': 12,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
             'name': 'Font Size (pt)',
@@ -346,7 +346,7 @@ class CustomModule(AbstractFunction):
         self.timer_loop = time.time()
         self.line_sets = []
         self.current_line_set = 0
-        self.line_y_dimensions = [0, 8, 16, 24, 32, 40, 48, 56]
+        self.line_y_dimensions = [0, 16]
         self.pad = -2
 
         # Initialize custom options
@@ -392,7 +392,7 @@ class CustomModule(AbstractFunction):
                 "i2c_bus": self.i2c_bus,
                 "x_characters": self.characters_x,
                 "line_y_dimensions": self.line_y_dimensions,
-                "lcd_type": "128x64_pioled_circuit_python",
+                "lcd_type": "128x32_pioled_circuit_python",
                 "font_size": self.font_size
             }
 
@@ -505,11 +505,7 @@ class CustomModule(AbstractFunction):
             lines_display[0],
             lines_display[1],
             lines_display[2],
-            lines_display[3],
-            lines_display[4],
-            lines_display[5],
-            lines_display[6],
-            lines_display[7])
+            lines_display[3])
 
     def stop_function(self):
         self.device.lcd_write_lines(